requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Document Processing & OCR
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3
PyPDF2==3.0.1
python-multipart==0.0.6

# Natural Language Processing
spacy==3.7.2
transformers==4.36.0
torch==2.1.1
sentence-transformers==2.2.2

# LLM Integration
openai==1.3.7
langchain==0.0.352
langchain-openai==0.0.2
langchain-community==0.0.6

# CrewAI
crewai==0.3.0
crewai-tools==0.4.0

# Model Context Protocol
mcp==0.3.0
jsonschema==4.20.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
httpx==0.25.2
aiofiles==23.2.1

# Background Tasks
celery==5.3.4
redis==5.0.1

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# Email
emails==0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
factory-boy==3.3.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Environment
python-dotenv==1.0.0

# Utilities
click==8.1.7
rich==13.7.0
jinja2==3.1.2